Fix UIP_LLADDR_LEN wrt 802.15.4 short addresses.
In order to get the size of the 802.15.4 addresses, the existing
code sometimes uses sizeof(rimeaddr_t) or RIMEADDR_SIZE, and
sometimes uses UIP_LLADDR_LEN. RIMEADDR_CONF_SIZE may be set to 2
in order to use short addresses, but UIP_LLADDR_LEN was always 8
(long address), which was causing various bugs.
This fix makes the 2 values equivalent.